[FEAT] BOJ11382_문제해결

+++ baek_joon/src/DAY04/BOJ11382.java
@@ -9,10 +9,9 @@ public class BOJ11382 {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
 
-        int a = sc.nextInt();
-        int b = sc.nextInt();
-        int c = sc.nextInt();
-
+        long a = sc.nextLong();
+        long b = sc.nextLong();
+        long c = sc.nextLong();
         System.out.println(a + b + c);
     }
 }
